+++ flutter/lib/src/integrations/native_app_start_integration.dart
@@ -1,7 +1,4 @@
-import 'dart:async';
 import 'dart:ui';
-
-import 'package:meta/meta.dart';
 
 import '../../sentry_flutter.dart';
 import '../frame_callback_handler.dart';
@@ -15,20 +12,7 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
 
   final FrameCallbackHandler _frameCallbackHandler;
   final NativeAppStartHandler _nativeAppStartHandler;
-  DateTime? _appStartEnd;
 
-  /// This timestamp marks the end of app startup. Either set by calling
-  /// [SentryFlutter.setAppStartEnd]. The [SentryFlutterOptions.autoAppStart]
-  /// option needs to be false.
-  @internal
-  set appStartEnd(DateTime appStartEnd) {
-    _appStartEnd = appStartEnd;
-    if (!_appStartEndCompleter.isCompleted) {
-      _appStartEndCompleter.complete();
-    }
-  }
-
-  final Completer<void> _appStartEndCompleter = Completer<void>();
   bool _allowProcessing = true;
 
   @override
@@ -42,26 +26,14 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
       _allowProcessing = false;
 
       try {
-        DateTime? appStartEnd;
-        if (options.autoAppStart) {
-          // ignore: invalid_use_of_internal_member
-          appStartEnd = DateTime.fromMicrosecondsSinceEpoch(timings.first
-              .timestampInMicroseconds(FramePhase.rasterFinishWallTime));
-        } else if (_appStartEnd == null) {
-          await _appStartEndCompleter.future.timeout(
-            const Duration(seconds: 10),
-          );
-          appStartEnd = _appStartEnd;
-        } else {
-          appStartEnd = null;
-        }
-        if (appStartEnd != null) {
-          await _nativeAppStartHandler.call(
-            hub,
-            options,
-            appStartEnd: appStartEnd,
-          );
-        }
+        // ignore: invalid_use_of_internal_member
+        final appStartEnd = DateTime.fromMicrosecondsSinceEpoch(timings.first
+            .timestampInMicroseconds(FramePhase.rasterFinishWallTime));
+        await _nativeAppStartHandler.call(
+          hub,
+          options,
+          appStartEnd: appStartEnd,
+        );
       } catch (exception, stackTrace) {
         options.logger(
           SentryLevel.error,